move all fixtures into conftest.py
