+++ tests/test_yaesm/test_timeframe.py
@@ -3,105 +3,6 @@ import random
 
 from yaesm.timeframe import Timeframe, FiveMinuteTimeframe, HourlyTimeframe, DailyTimeframe, WeeklyTimeframe, MonthlyTimeframe, YearlyTimeframe
 
-@pytest.fixture
-def random_timeframe_generator(random_timeframe_times_generator, random_timeframe_minutes_generator, random_timeframe_weekdays_generator, random_timeframe_monthdays_generator, random_timeframe_yeardays_generator):
-    """Fixture for generating a random Timeframe."""
-    def generator(tframe_type=None, keep=None, minutes=None, times=None, weekdays=None, monthdays=None, yeardays=None) -> Timeframe:
-        tframe_type = random.choice(Timeframe.valid_tframe_types()) if tframe_type is None else tframe_type
-        keep        = random.randint(0,10) if keep is None else keep
-        minutes     = random_timeframe_minutes_generator(random.randint(0,5)) if minutes is None else minutes
-        times       = random_timeframe_times_generator(random.randint(0,5)) if times is None else times
-        weekdays    = random_timeframe_weekdays_generator(random.randint(0,3)) if weekdays is None else weekdays
-        monthdays   = random_timeframe_monthdays_generator(random.randint(0,3)) if monthdays is None else monthdays
-        yeardays    = random_timeframe_yeardays_generator(random.randint(0,3)) if yeardays is None else yeardays
-        if tframe_type == FiveMinuteTimeframe:
-            return FiveMinuteTimeframe(keep)
-        elif tframe_type == HourlyTimeframe:
-            return HourlyTimeframe(keep, minutes)
-        elif tframe_type == DailyTimeframe:
-            return DailyTimeframe(keep, times)
-        elif tframe_type == WeeklyTimeframe:
-            return WeeklyTimeframe(keep, times, weekdays)
-        elif tframe_type == MonthlyTimeframe:
-            return MonthlyTimeframe(keep, times, monthdays)
-        elif tframe_type == YearlyTimeframe:
-            return YearlyTimeframe(keep, times, yeardays)
-    return generator
-
-@pytest.fixture
-def random_timeframes_generator(random_timeframe_generator):
-    """Fixture for generating a list of random Timeframes. See the fixture
-    random_timeframe_generator for more details."""
-    def generator(num=3, **kwargs):
-        timeframes = []
-        tframe_types = random.sample(Timframe.valid_tframe_types, k=num)
-        for tframe_type in tframe_types:
-            timeframes.append(random_timeframe_generator(tframe_type=tframe_type, **kwargs))
-        return timeframes
-    return generator
-
-@pytest.fixture
-def random_timeframe_minutes_generator():
-    """Fixture to generate a list of random minutes."""
-    def generator(num=3):
-        minutes = []
-        for _ in range(num):
-            minute = None
-            while minute is None or minute in minutes:
-                minute = random.randint(0,59)
-            minutes.append(time)
-        return minutes
-    return generator
-
-@pytest.fixture
-def random_timeframe_timespecs_generator():
-    """Fixture to generate a list of random timeframe timespecs."""
-    def generator(num=3):
-        timepsecs = []
-        for _ in range(num):
-            timespec = None
-            while timespec is None or timespec in timespecs:
-                hour = str(random.randint(0,23)).zfill(2)
-                minute = str(random.randint(0,59)).zfill(2)
-                timespec = f"{hour}:{minute}"
-            timespecs.append(timespec)
-        return timespecs
-    return generator
-
-@pytest.fixture
-def random_timeframe_times_generator(random_timeframe_timespecs_generator):
-    """Fixture to generate a list of random timeframe times."""
-    def generator(num=3):
-        timespecs = random_timeframe_timespecs_generator(num=num)
-        times = list(map(lambda x: Timeframe.timespec_to_time(x), timespecs))
-        return times
-    return generator
-
-@pytest.fixture
-def random_timeframe_weekdays_generator():
-    """Fixture to generate a list of random timeframe weekdays."""
-    def generator(num=3):
-        weekdays = random.sample(["monday","tuesday","wednesday","thursday","friday","saturday","sunday"], k=num)
-        return weekdays
-    return generator
-
-@pytest.fixture
-def random_timeframe_monthdays_generator():
-    """Fixture to generate a list or random Timeframe monthdays."""
-    def generator(num=3):
-        monthdays = random.sample(list(range(1,32)), k=num)
-        return monthdays
-    return generator
-
-@pytest.fixture
-def random_timeframe_yeardays_generator():
-    """Fixture to generate a list of random Timeframe yeardays."""
-    def generator(num=3, include_leap=False):
-        max_day = 366 if include_leap else 365
-        yeardays = random.sample(list(range(1,max_day+1)), k=num)
-        return yeardays
-    return generator
-
 def test_FiveMinuteTimeframe():
     assert FiveMinuteTimeframe.required_config_settings() == ["5minute_keep"]
     tf = FiveMinuteTimeframe(10)
